test(Mutiny): add integration test
- move existing tests from return-response to own module
