+++ client/integration-tests/mutiny-return-response/src/main/openapi/mutiny-multi-return-response-false-string-simple-openapi.yaml
