+++ client/integration-tests/mutiny-return-response/src/main/openapi/mutiny-multi-return-response-false-void-simple-openapi.yaml
