+++ client/integration-tests/mutiny-return-response/src/main/openapi/mutiny-multi-return-response-true-string-simple-openapi.yaml
