+++ client/integration-tests/mutiny-return-response/src/main/openapi/mutiny-return-response-false-string-simple-openapi.yaml
