+++ client/integration-tests/mutiny-return-response/src/main/openapi/mutiny-return-response-true-string-simple-openapi.yaml
