+++ client/integration-tests/mutiny-return-response/src/main/resources/application.properties
@@ -1,0 +1,36 @@
+quarkus.openapi-generator.codegen.spec.mutiny_return_response_false_string_simple_openapi_yaml.base-package = org.acme.openapi
+quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_false_string_simple_openapi_yaml.base-package = org.acme.openapi
+quarkus.openapi-generator.codegen.spec.mutiny_return_response_false_void_simple_openapi_yaml.base-package = org.acme.openapi
+quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_false_void_simple_openapi_yaml.base-package = org.acme.openapi
+quarkus.openapi-generator.codegen.spec.mutiny_return_response_true_string_simple_openapi_yaml.base-package = org.acme.openapi
+quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_true_string_simple_openapi_yaml.base-package = org.acme.openapi
+quarkus.openapi-generator.codegen.spec.mutiny_return_response_true_void_simple_openapi_yaml.base-package = org.acme.openapi
+quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_true_void_simple_openapi_yaml.base-package = org.acme.openapi
+
+quarkus.openapi-generator.codegen.spec.mutiny_return_response_false_string_simple_openapi_yaml.mutiny = true
+quarkus.openapi-generator.codegen.spec.mutiny_return_response_false_string_simple_openapi_yaml.mutiny.return-response = false
+
+quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_false_string_simple_openapi_yaml.mutiny = true
+quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_false_string_simple_openapi_yaml.mutiny.return-response = false
+quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_false_string_simple_openapi_yaml.mutiny.operation-ids.hello = Multi
+
+quarkus.openapi-generator.codegen.spec.mutiny_return_response_false_void_simple_openapi_yaml.mutiny = true
+quarkus.openapi-generator.codegen.spec.mutiny_return_response_false_void_simple_openapi_yaml.mutiny.return-response = false
+
+quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_false_void_simple_openapi_yaml.mutiny = true
+quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_false_void_simple_openapi_yaml.mutiny.return-response = false
+quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_false_void_simple_openapi_yaml.mutiny.operation-ids.hello = Multi
+
+quarkus.openapi-generator.codegen.spec.mutiny_return_response_true_string_simple_openapi_yaml.mutiny = true
+quarkus.openapi-generator.codegen.spec.mutiny_return_response_true_string_simple_openapi_yaml.mutiny.return-response = true
+
+quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_true_string_simple_openapi_yaml.mutiny = true
+quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_true_string_simple_openapi_yaml.mutiny.return-response = true
+quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_true_string_simple_openapi_yaml.mutiny.operation-ids.hello = Multi
+
+quarkus.openapi-generator.codegen.spec.mutiny_return_response_true_void_simple_openapi_yaml.mutiny = true
+quarkus.openapi-generator.codegen.spec.mutiny_return_response_true_void_simple_openapi_yaml.mutiny.return-response = true
+
+quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_true_void_simple_openapi_yaml.mutiny = true
+quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_true_void_simple_openapi_yaml.mutiny.return-response = true
+quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_true_void_simple_openapi_yaml.mutiny.operation-ids.hello = Multi
