+++ client/integration-tests/pom.xml
@@ -27,6 +27,7 @@
     <module>include</module>
     <module>multipart-request</module>
     <module>mutiny</module>
+    <module>mutiny-return-response</module>
     <module>open-api-normalizer</module>
     <module>part-filename</module>
     <module>return-response</module>
