+++ client/integration-tests/return-response/src/main/resources/application.properties
@@ -12,34 +12,3 @@ quarkus.openapi-generator.codegen.spec.return_response_false_string_simple_opena
 quarkus.openapi-generator.codegen.spec.return_response_false_void_simple_openapi_yaml.mutiny = false
 quarkus.openapi-generator.codegen.spec.return_response_true_string_simple_openapi_yaml.mutiny = false
 quarkus.openapi-generator.codegen.spec.return_response_true_void_simple_openapi_yaml.mutiny = false
-
-quarkus.openapi-generator.codegen.spec.mutiny_return_response_false_string_simple_openapi_yaml.base-package = ${quarkus.openapi-generator.codegen.spec.return_response_false_string_simple_openapi_yaml.base-package}
-quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_false_string_simple_openapi_yaml.base-package = ${quarkus.openapi-generator.codegen.spec.return_response_false_string_simple_openapi_yaml.base-package}
-quarkus.openapi-generator.codegen.spec.mutiny_return_response_false_void_simple_openapi_yaml.base-package = ${quarkus.openapi-generator.codegen.spec.return_response_false_void_simple_openapi_yaml.base-package}
-quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_false_void_simple_openapi_yaml.base-package = ${quarkus.openapi-generator.codegen.spec.return_response_false_void_simple_openapi_yaml.base-package}
-quarkus.openapi-generator.codegen.spec.mutiny_return_response_true_string_simple_openapi_yaml.base-package = ${quarkus.openapi-generator.codegen.spec.return_response_true_string_simple_openapi_yaml.base-package}
-quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_true_string_simple_openapi_yaml.base-package = ${quarkus.openapi-generator.codegen.spec.return_response_true_string_simple_openapi_yaml.base-package}
-quarkus.openapi-generator.codegen.spec.mutiny_return_response_true_void_simple_openapi_yaml.base-package = ${quarkus.openapi-generator.codegen.spec.return_response_true_void_simple_openapi_yaml.base-package}
-quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_true_void_simple_openapi_yaml.base-package = ${quarkus.openapi-generator.codegen.spec.return_response_true_void_simple_openapi_yaml.base-package}
-
-quarkus.openapi-generator.codegen.spec.mutiny_return_response_false_string_simple_openapi_yaml.return-response = ${quarkus.openapi-generator.codegen.spec.return_response_false_string_simple_openapi_yaml.return-response}
-quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_false_string_simple_openapi_yaml.return-response = ${quarkus.openapi-generator.codegen.spec.multi_return_response_false_string_simple_openapi_yaml.return-response}
-quarkus.openapi-generator.codegen.spec.mutiny_return_response_false_void_simple_openapi_yaml.return-response = ${quarkus.openapi-generator.codegen.spec.return_response_false_void_simple_openapi_yaml.return-response}
-quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_false_void_simple_openapi_yaml.return-response = ${quarkus.openapi-generator.codegen.spec.return_response_false_void_simple_openapi_yaml.return-response}
-quarkus.openapi-generator.codegen.spec.mutiny_return_response_true_string_simple_openapi_yaml.mutiny.return-response = ${quarkus.openapi-generator.codegen.spec.return_response_true_string_simple_openapi_yaml.return-response}
-quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_true_string_simple_openapi_yaml.mutiny.return-response = ${quarkus.openapi-generator.codegen.spec.return_response_true_string_simple_openapi_yaml.return-response}
-quarkus.openapi-generator.codegen.spec.mutiny_return_response_true_void_simple_openapi_yaml.return-response = ${quarkus.openapi-generator.codegen.spec.return_response_true_void_simple_openapi_yaml.return-response}
-quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_true_void_simple_openapi_yaml.return-response = ${quarkus.openapi-generator.codegen.spec.return_response_true_void_simple_openapi_yaml.return-response}
-
-quarkus.openapi-generator.codegen.spec.mutiny_return_response_false_string_simple_openapi_yaml.mutiny = true
-quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_false_string_simple_openapi_yaml.mutiny = true
-quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_false_string_simple_openapi_yaml.mutiny.operation-ids.hello = Multi
-quarkus.openapi-generator.codegen.spec.mutiny_return_response_false_void_simple_openapi_yaml.mutiny = true
-quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_false_void_simple_openapi_yaml.mutiny = true
-quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_false_void_simple_openapi_yaml.mutiny.operation-ids.hello = Multi
-quarkus.openapi-generator.codegen.spec.mutiny_return_response_true_string_simple_openapi_yaml.mutiny = true
-quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_true_string_simple_openapi_yaml.mutiny = true
-quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_true_string_simple_openapi_yaml.mutiny.operation-ids.hello = Multi
-quarkus.openapi-generator.codegen.spec.mutiny_return_response_true_void_simple_openapi_yaml.mutiny = true
-quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_true_void_simple_openapi_yaml.mutiny = true
-quarkus.openapi-generator.codegen.spec.mutiny_multi_return_response_true_void_simple_openapi_yaml.mutiny.operation-ids.hello = Multi
